+++ src/lib/core/logging.h
@@ -27,7 +27,7 @@ void log_bind_uart(UartState_t *u);
 
 // Emit to log system. Not really meant to be called directly, but through the
 // LOG() macro.
-void log_write(const void *buf, int len);
+void log_write(const void *buf, size_t len);
 
 // wait until log is drained
 void log_flush();
