+++ src/lib/periph/uart/uart.c
@@ -86,7 +86,7 @@ static void _get_next_data(uint8_t uart_id, void *user_data,
   }
 }
 
-void uart_write(UartState_t *u, const void *buf, uint16_t len) {
+void uart_write(UartState_t *u, const void *buf, size_t len) {
   const char *c = (char *)buf;
   assert(u->initted);
 
@@ -95,7 +95,8 @@ void uart_write(UartState_t *u, const void *buf, uint16_t len) {
 
     // In critical section: add as much new data to the send queue as will fit.
     rulos_irq_state_t old_interrupts = hal_start_atomic();
-    uint16_t write_size = min(len, CharQueue_free_space(&u->tx_queue.q));
+    uint16_t write_size =
+      min((size_t)len, CharQueue_free_space(&u->tx_queue.q));
 
     if (write_size == 0) {
       // If there's more data that did not fit in the queue, block until there's
